Reorganize MSE::reduce
