+++ src/expr/src/scalar/reduce/binary.rs
@@ -1,0 +1,323 @@
+// Copyright Materialize, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+//! Post-order rewrites for `CallBinary` nodes.
+
+use std::mem;
+
+use itertools::Itertools;
+use mz_pgtz::timezone::{Timezone, TimezoneSpec};
+use mz_repr::adt::datetime::DateTimeUnits;
+use mz_repr::adt::regex::Regex;
+use mz_repr::{Datum, ReprColumnType, ReprScalarType, RowArena};
+
+use crate::EvalError;
+use crate::MirScalarExpr;
+use crate::scalar::func::format::DateTimeFormat;
+use crate::scalar::func::variadic::And;
+use crate::scalar::func::{self, BinaryFunc, UnaryFunc, VariadicFunc, parse_timezone};
+use crate::scalar::like_pattern;
+
+pub(super) fn reduce_call_binary(
+    e: &mut MirScalarExpr,
+    column_types: &[ReprColumnType],
+    temp_storage: &RowArena,
+) {
+    let MirScalarExpr::CallBinary { func, expr1, expr2 } = e else {
+        unreachable!()
+    };
+
+    // Fold/propagate literal-shaped operands first; precompiles below assume
+    // these have already fired.
+    if expr1.is_literal() && expr2.is_literal() {
+        *e = MirScalarExpr::literal(e.eval(&[], temp_storage), e.typ(column_types).scalar_type);
+        return;
+    }
+    if (expr1.is_literal_null() || expr2.is_literal_null()) && func.propagates_nulls() {
+        *e = MirScalarExpr::literal_null(e.typ(column_types).scalar_type);
+        return;
+    }
+    if let Some(err) = expr1.as_literal_err() {
+        *e = MirScalarExpr::literal(Err(err.clone()), e.typ(column_types).scalar_type);
+        return;
+    }
+    if let Some(err) = expr2.as_literal_err() {
+        *e = MirScalarExpr::literal(Err(err.clone()), e.typ(column_types).scalar_type);
+        return;
+    }
+
+    // Per-function dispatch. Each precompile fires only if its literal-shaped
+    // argument is present; otherwise the call falls through unchanged.
+    match func {
+        BinaryFunc::IsLikeMatchCaseInsensitive(_) if expr2.is_literal() => {
+            precompile_is_like(e, column_types, true);
+        }
+        BinaryFunc::IsLikeMatchCaseSensitive(_) if expr2.is_literal() => {
+            precompile_is_like(e, column_types, false);
+        }
+        BinaryFunc::IsRegexpMatchCaseSensitive(_) | BinaryFunc::IsRegexpMatchCaseInsensitive(_) => {
+            let case_insensitive = matches!(func, BinaryFunc::IsRegexpMatchCaseInsensitive(_));
+            if let MirScalarExpr::Literal(Ok(row), _) = &**expr2 {
+                *e = match Regex::new(row.unpack_first().unwrap_str(), case_insensitive) {
+                    Ok(regex) => expr1
+                        .take()
+                        .call_unary(UnaryFunc::IsRegexpMatch(func::IsRegexpMatch(regex))),
+                    Err(err) => {
+                        MirScalarExpr::literal(Err(err.into()), e.typ(column_types).scalar_type)
+                    }
+                };
+            }
+        }
+        BinaryFunc::ExtractInterval(_) if expr1.is_literal() => {
+            precompile_date_units(e, column_types, |u| {
+                UnaryFunc::ExtractInterval(func::ExtractInterval(u))
+            })
+        }
+        BinaryFunc::ExtractTime(_) if expr1.is_literal() => {
+            precompile_date_units(e, column_types, |u| {
+                UnaryFunc::ExtractTime(func::ExtractTime(u))
+            })
+        }
+        BinaryFunc::ExtractTimestamp(_) if expr1.is_literal() => {
+            precompile_date_units(e, column_types, |u| {
+                UnaryFunc::ExtractTimestamp(func::ExtractTimestamp(u))
+            })
+        }
+        BinaryFunc::ExtractTimestampTz(_) if expr1.is_literal() => {
+            precompile_date_units(e, column_types, |u| {
+                UnaryFunc::ExtractTimestampTz(func::ExtractTimestampTz(u))
+            })
+        }
+        BinaryFunc::ExtractDate(_) if expr1.is_literal() => {
+            precompile_date_units(e, column_types, |u| {
+                UnaryFunc::ExtractDate(func::ExtractDate(u))
+            })
+        }
+        BinaryFunc::DatePartInterval(_) if expr1.is_literal() => {
+            precompile_date_units(e, column_types, |u| {
+                UnaryFunc::DatePartInterval(func::DatePartInterval(u))
+            })
+        }
+        BinaryFunc::DatePartTime(_) if expr1.is_literal() => {
+            precompile_date_units(e, column_types, |u| {
+                UnaryFunc::DatePartTime(func::DatePartTime(u))
+            })
+        }
+        BinaryFunc::DatePartTimestamp(_) if expr1.is_literal() => {
+            precompile_date_units(e, column_types, |u| {
+                UnaryFunc::DatePartTimestamp(func::DatePartTimestamp(u))
+            })
+        }
+        BinaryFunc::DatePartTimestampTz(_) if expr1.is_literal() => {
+            precompile_date_units(e, column_types, |u| {
+                UnaryFunc::DatePartTimestampTz(func::DatePartTimestampTz(u))
+            })
+        }
+        BinaryFunc::DateTruncTimestamp(_) if expr1.is_literal() => {
+            precompile_date_units(e, column_types, |u| {
+                UnaryFunc::DateTruncTimestamp(func::DateTruncTimestamp(u))
+            })
+        }
+        BinaryFunc::DateTruncTimestampTz(_) if expr1.is_literal() => {
+            precompile_date_units(e, column_types, |u| {
+                UnaryFunc::DateTruncTimestampTz(func::DateTruncTimestampTz(u))
+            })
+        }
+        BinaryFunc::TimezoneTimestampBinary(_) if expr1.is_literal() => {
+            // If the timezone argument is a literal, and we're applying the
+            // function on many rows, we want to parse it once and embed it
+            // into the UnaryFunc enum. The memory footprint of Timezone is
+            // small (8 bytes).
+            precompile_timezone(e, column_types, |tz| {
+                UnaryFunc::TimezoneTimestamp(func::TimezoneTimestamp(tz))
+            });
+        }
+        BinaryFunc::TimezoneTimestampTzBinary(_) if expr1.is_literal() => {
+            precompile_timezone(e, column_types, |tz| {
+                UnaryFunc::TimezoneTimestampTz(func::TimezoneTimestampTz(tz))
+            });
+        }
+        BinaryFunc::ToCharTimestamp(_) if expr2.is_literal() => {
+            precompile_to_char(e, |format_string, format| {
+                UnaryFunc::ToCharTimestamp(func::ToCharTimestamp {
+                    format_string,
+                    format,
+                })
+            });
+        }
+        BinaryFunc::ToCharTimestampTz(_) if expr2.is_literal() => {
+            precompile_to_char(e, |format_string, format| {
+                UnaryFunc::ToCharTimestampTz(func::ToCharTimestampTz {
+                    format_string,
+                    format,
+                })
+            });
+        }
+        BinaryFunc::Eq(_) | BinaryFunc::NotEq(_) if expr2 < expr1 => {
+            // Canonically order elements so that deduplication works better.
+            // Also, the `Literal([c1, c2]) = record_create(e1, e2)` matching
+            // below relies on this canonical ordering.
+            mem::swap(expr1, expr2);
+        }
+        _ => reduce_call_binary_eq_record(e),
+    }
+}
+
+/// Decomposes equality with `RecordCreate` on the right-hand side.
+///
+/// Handles two cases:
+/// - `Literal([c1, ...]) = record_create(e1, ...)` → `c1 = e1 AND ...`
+/// - `record_create(a1, ...) = record_create(b1, ...)` → `a1 = b1 AND ...`
+///
+/// `MapFilterProject::literal_constraints` relies on the first transform,
+/// because `(e1,e2) IN ((1,2))` is desugared using `record_create`.
+fn reduce_call_binary_eq_record(e: &mut MirScalarExpr) {
+    let MirScalarExpr::CallBinary { func, expr1, expr2 } = e else {
+        unreachable!()
+    };
+
+    match (&*func, &**expr1, &**expr2) {
+        (
+            BinaryFunc::Eq(_),
+            MirScalarExpr::Literal(
+                Ok(lit_row),
+                ReprColumnType {
+                    scalar_type:
+                        ReprScalarType::Record {
+                            fields: field_types,
+                            ..
+                        },
+                    ..
+                },
+            ),
+            MirScalarExpr::CallVariadic {
+                func: VariadicFunc::RecordCreate(..),
+                exprs: rec_create_args,
+            },
+        ) => {
+            if let Datum::List(datum_list) = lit_row.unpack_first() {
+                *e = MirScalarExpr::call_variadic(
+                    And,
+                    datum_list
+                        .iter()
+                        .zip_eq(field_types)
+                        .zip_eq(rec_create_args)
+                        .map(|((d, typ), a)| {
+                            MirScalarExpr::literal_ok(d, typ.scalar_type.clone())
+                                .call_binary(a.clone(), func::Eq)
+                        })
+                        .collect(),
+                );
+            }
+        }
+        (
+            BinaryFunc::Eq(_),
+            MirScalarExpr::CallVariadic {
+                func: VariadicFunc::RecordCreate(..),
+                exprs: rec_create_args1,
+            },
+            MirScalarExpr::CallVariadic {
+                func: VariadicFunc::RecordCreate(..),
+                exprs: rec_create_args2,
+            },
+        ) => {
+            // Note that there is a similar decomposition in
+            // `mz_sql::plan::transform_ast::Desugarer`, but that is earlier in
+            // the pipeline than the compilation of IN lists to `record_create`.
+            *e = MirScalarExpr::call_variadic(
+                And,
+                rec_create_args1
+                    .into_iter()
+                    .zip_eq(rec_create_args2)
+                    .map(|(a, b)| a.clone().call_binary(b.clone(), func::Eq))
+                    .collect(),
+            );
+        }
+        _ => {}
+    }
+}
+
+/// Specializes a binary date/time function call whose units argument is a
+/// literal string. Produces either a unary call with the parsed units baked
+/// in, or a literal `UnknownUnits` error.
+fn precompile_date_units<F>(e: &mut MirScalarExpr, column_types: &[ReprColumnType], build_unary: F)
+where
+    F: FnOnce(DateTimeUnits) -> UnaryFunc,
+{
+    let MirScalarExpr::CallBinary { expr1, expr2, .. } = e else {
+        unreachable!()
+    };
+    let units_str = expr1.as_literal_str().unwrap();
+    *e = match units_str.parse::<DateTimeUnits>() {
+        Ok(units) => MirScalarExpr::CallUnary {
+            func: build_unary(units),
+            expr: Box::new(expr2.take()),
+        },
+        Err(_) => MirScalarExpr::literal(
+            Err(EvalError::UnknownUnits(units_str.into())),
+            e.typ(column_types).scalar_type,
+        ),
+    };
+}
+
+/// Specializes a binary timezone-applying function call whose timezone
+/// argument is a literal. Produces either a unary call with the parsed
+/// timezone baked in, or a literal error.
+fn precompile_timezone<F>(e: &mut MirScalarExpr, column_types: &[ReprColumnType], build_unary: F)
+where
+    F: FnOnce(Timezone) -> UnaryFunc,
+{
+    let MirScalarExpr::CallBinary { expr1, expr2, .. } = e else {
+        unreachable!()
+    };
+    let tz_str = expr1.as_literal_str().unwrap();
+    *e = match parse_timezone(tz_str, TimezoneSpec::Posix) {
+        Ok(tz) => MirScalarExpr::CallUnary {
+            func: build_unary(tz),
+            expr: Box::new(expr2.take()),
+        },
+        Err(err) => MirScalarExpr::literal(Err(err), e.typ(column_types).scalar_type),
+    };
+}
+
+/// Specializes a `to_char_*` binary call whose format-string argument is a
+/// literal, by compiling the format into the unary form.
+fn precompile_to_char<F>(e: &mut MirScalarExpr, build_unary: F)
+where
+    F: FnOnce(String, DateTimeFormat) -> UnaryFunc,
+{
+    let MirScalarExpr::CallBinary { expr1, expr2, .. } = e else {
+        unreachable!()
+    };
+    let format_str = expr2.as_literal_str().unwrap().to_owned();
+    let compiled = DateTimeFormat::compile(&format_str);
+    *e = MirScalarExpr::CallUnary {
+        func: build_unary(format_str, compiled),
+        expr: Box::new(expr1.take()),
+    };
+}
+
+/// Specializes an `IsLikeMatch{CaseSensitive,CaseInsensitive}` binary call
+/// whose pattern argument is a literal, by precompiling the matcher.
+fn precompile_is_like(
+    e: &mut MirScalarExpr,
+    column_types: &[ReprColumnType],
+    case_insensitive: bool,
+) {
+    let MirScalarExpr::CallBinary { expr1, expr2, .. } = e else {
+        unreachable!()
+    };
+    let pattern = expr2.as_literal_str().unwrap();
+    *e = match like_pattern::compile(pattern, case_insensitive) {
+        Ok(matcher) => expr1
+            .take()
+            .call_unary(UnaryFunc::IsLikeMatch(func::IsLikeMatch(matcher))),
+        Err(err) => MirScalarExpr::literal(Err(err), e.typ(column_types).scalar_type),
+    };
+}
